+++ src/index.ts
@@ -5,5 +5,7 @@ export * from "./engine/interactive";
 export * from "./engine/envWriter";
 export * from "./engine/envGenerator";
 export * from "./engine/codeScanner";
+export * from "./engine/frameworkChecker";
+export * from "./reporter/sarifReporter";
 export * from "./config/loadConfig";
 export * from "./types";